Add new fields

+++ server/backend-app/src/main/java/com/sawwere/titlecounter/backend/app/storage/entity/GameEntry.java
@@ -31,7 +31,7 @@ public class GameEntry {
 
     @Min(1)
     @Max(10)
-    private Long score;
+    private Integer score;
 
     @NotBlank
     @Column(nullable = false)
